# openapi.yaml
openapi: 3.0.3
info:
  title: Panda Market API
  version: 1.0.0
servers:
  - url: http://localhost:3001
...
paths:
...
  /users/me:
    get:
      tags: [User]
      description: 로그인한 사용자 정보 조회
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    patch:
      tags: [User]
      description: 로그인한 사용자 프로필 수정
      operationId: updateMe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  nullable: true
      responses:
        '200':
          description: 수정 성공
  /users/me/password:
    patch:
      tags: [User]
      description: 로그인한 사용자 비밀번호 수정
      operationId: updateMyPassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password, passwordConfirmation, currentPassword]
              properties:
                password: { type: string }
                passwordConfirmation: { type: string }
                currentPassword: { type: string }
      responses:
        '200':
          description: 수정 성공
  /users/me/products:
    get:
      tags: [User]
      description: 내가 등록한 상품 목록 조회
      operationId: getMyProducts
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Keyword'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProductListResponse'
  /users/me/favorites:
    get:
      tags: [User]
      description: 내가 좋아요한 상품 목록 조회
      operationId: getMyFavoriteProducts
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Keyword'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProductListResponse'
...
  /products:
    post:
      tags:
        - Product
      description: 상품 생성 API
      operationId: createProduct
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Validation failed

    get:
      tags:
        - Product
      description: 상품 목록 조회 API
      operationId: getProductList
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
          example: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 12
          example: 10
        - name: orderBy
          in: query
          required: false
          schema:
            type: string
            enum:
              - recent
              - favorite
          example: recent
        - name: keyword
          in: query
          required: false
          schema:
            type: string
          example: "검색어"
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProductListResponse'
  /products/{productId}:
    get:
      tags:
        - Product
      description: 상품 조회 API
      operationId: getProduct
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not found

    patch:
      tags:
        - Product
      description: "상품 수정 API"
      operationId: updateProduct
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Validation failed
        '404':
          description: Not found

    delete:
      tags:
        - Product
      description: "상품 삭제 API"
      operationId: deleteProduct
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '204':
          description: 성공
        '404':
          description: Not found
  /products/{productId}/favorite:
    post:
      tags: [Product]
      description: 상품 좋아요
      operationId: favoriteProduct
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '201':
          description: 좋아요 성공
    delete:
      tags: [Product]
      description: 상품 좋아요 취소
      operationId: unfavoriteProduct
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '201':
          description: 좋아요 취소 성공
  /products/{productId}/comments:
    post:
      tags:
        - Product
      description: 댓글 생성 API
      operationId: createProductComment
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductCommentRequest'
      responses:
        '201':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Validation failed
        '404':
          description: Not found

    get:
      tags:
        - Product
      description: 댓글 목록 조회 API
      operationId: getProductCommentList
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
          example: 1
        - name: cursor
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
          example: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 10
          example: 10
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProductCommentListResponse'
        '404':
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 10
    Keyword:
      name: keyword
      in: query
      required: false
      schema:
        type: string
...
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        nickname:
          type: string
        image:
          type: string
          nullable: true
...
    CreateProductRequest:
      type: object
      required:
        - name
        - description
        - price
        - tags
        - images
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 10
          example: "상품 이름"
        description:
          type: string
          minLength: 10
          maxLength: 100
          example: "상품 설명"
        price:
          type: integer
          minimum: 0
          example: 10000
        tags:
          type: array
          minItems: 1
          items:
            type: string
            minLength: 1
            maxLength: 5
          example: ["태그1", "태그2"]
        images:
          type: array
          maxItems: 3
          items:
            type: string
          example: ["http://localhost:13000/static/images/sample-image.png"]
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 10
          example: "상품 이름"
        description:
          type: string
          minLength: 10
          maxLength: 100
          example: "상품 설명"
        price:
          type: integer
          minimum: 0
          example: 10000
        tags:
          type: array
          minItems: 1
          items:
            type: string
            minLength: 1
            maxLength: 5
          example: ["태그1", "태그2"]
        images:
          type: array
          maxItems: 3
          items:
            type: string
          example: ["http://localhost:13000/static/images/sample-image.png"]
    CreateProductCommentRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          example: "댓글 내용"
    GetProductCommentListResponse:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        nextCursor:
          type: integer
          nullable: true
          example: 10
...
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "상품 이름"
        description:
          type: string
          example: "상품 설명"
        price:
          type: integer
          example: 10000
        tags:
          type: array
          items:
            type: string
          example: ["태그1", "태그2"]
        images:
          type: array
          items:
            type: string
          example: ["http://localhost:13000/static/images/sample-image.png"]
        createdAt:
          type: string
          format: datetime
          example: "2024-07-15T00:00:00.000Z"
        favoriteCount:
          type: integer
          example: 3
        isLiked:
          type: boolean
          example: false
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
    GetProductListResponse:
      type: object
      required: [totalCount, list]
      properties:
        totalCount:
          type: integer
          example: 20
        list:
          type: array
          items:
            $ref: '#/components/schemas/Product'
    Comment:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: datetime
          example: "2024-07-15T00:00:00.000Z"